requirements.txt
python-dotenv~=1.0.0
aiogram~=3.0.0
requests~=2.31.0
gTTS~=2.2.4
pandas~=2.1.3
SQLAlchemy~=2.0.23